mend.ebnf: fix recursive rule generation.
The test.check gen-recursive generator takes a recursive generator
and a non-recursive generator as options. For directly recursive
grammar rules, a version of the rule without the recursion (smallest
optional part of the rule that recurses is removed) is used for the
non-recursive part of gen-recursive and the full recursive rule is
used for the recursive generator to recursive-gen.

- Update to test.check 0.9.1-SNAPSHOT
- Improve docstrings, comments.