Workaround geckodriver bugs: always have content.
geckodriver/firefox marionette mode have at least one bug in the
screenshot. The main one is that the screenshot code capture the
current content in the body rather than capturing the viewport. This
is an existing bug at:
https://bugzilla.mozilla.org/show_bug.cgi?id=1385706

A separate bug is that when the page is in fact empty and has no
actual content (margin 0px) but has a background color, the screenshot
command returns no data rather than a valid 0x0 PNG. This bug was
filed as: https://bugzilla.mozilla.org/show_bug.cgi?id=1492357

So to workaround those bugs we always generate a page with a single
'x' character as the first content within the body. This is a hack
that should be removed in the future because it may hide bugs that
happen when elements are the first thing on the page.