Add html-mangle, grammar caching. Fix prop/nonprop.
Move html-mangle and parse grammar caching from the wend module of
bartender to this project. This functionality is fairly generic and
more applicable to the html5/css3 grammar than to bartender.

Also, fix the prop/nonprop issue that's been worked around for a long
time. Turns out that on the right hand side the all prop
references were being translated to nonprop references which is why
the workaround of adding a bunch of nonprop -> prop references was
necessary. This should no longer be necessary.